feat: Added database insert functionality
Added database insert functionality code for streamers. However, the
code is currently out of place as a standalone and still needs to be
integrated into the scraping process. When integrating, need to remember
to implement insertion for stream events in to the stream table as well.
Lastly, implemented prefilling of category table using the categories
list already in the codebase. This prefilling only occurs when the
categories table does not already exist.