feat: Completed scraping and storage pipeline
Implemented full scraping and storage pipeline. Still contains a small
bug where the code crashes if a collected stream is now unavailable
(not meaning offline, but identical to if a streamer was banned and the
channel was removed from Twitch ... e.g. time machine error message).
Completed a full scrape of Overwatch and Dota 2 data which can be used
to benchmark learning agents.